currency double to decial, fix summary heading

+++ API/Models/BankDTOs/BankAccountCreateDTO.cs
@@ -6,8 +6,8 @@ public class BankAccountCreateDTO
     public string AppUserId { get; set; }
 
     public string HolderName { get; set; }
-    public double SavingsBalance { get; set; }
-    public double CheckingsBalance { get; set; }
-    public double TransactionAmount { get; set; }
+    public decimal SavingsBalance { get; set; }
+    public decimal CheckingsBalance { get; set; }
+    public decimal TransactionAmount { get; set; }
 }
